Create script SQL,exception change in service classes

+++ src/main/java/com/ITComponentsStore/Service/impl/ProductServiceImpl.java
@@ -1,5 +1,6 @@
 package com.ITComponentsStore.Service.impl;
 
+import com.ITComponentsStore.Entity.Category;
 import com.ITComponentsStore.Entity.Product;
 import com.ITComponentsStore.Exception.DataNotFoundException;
 import com.ITComponentsStore.Exception.InvalidDataException;
@@ -34,7 +35,7 @@ public class ProductServiceImpl implements ProductService {
         }
         Optional<Product> optionalProduct = productRepository.findById(id);
         if(optionalProduct.isEmpty()){
-            throw new DataNotFoundException("Error! The product does not exist! ");
+            throw new DataNotFoundException("Error! The product with id " + id+" does not exist! ");
         }
         else {
             return optionalProduct;
@@ -55,13 +56,15 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public void deleteProductById(Long id) {
+        Optional<Product> categoryOptional = productRepository.findById(id);
         if(id<0){
-            throw new InvalidDataException("Error! Your id" + id+ "it's not valid");
+            throw new InvalidDataException("Error! Your id" + id+ " it's not valid, pleas try again with id >=0");
         }
-        else{
-            productRepository.deleteById(id);
+        if(categoryOptional.isEmpty()){
+            throw new DataNotFoundException("Error Category with id " + id + " it's not present in database");
+        }
 
-        }
+        productRepository.deleteById(id);
 
     }
 }
